// src/webparts/hrContractsearch/components/IHrContractSearchProps.d.ts
export interface IHrContractSearchProps {
    description: string;
    isDarkTheme: boolean;
    environmentMessage: string;
    hasTeamsContext: boolean;
    userDisplayName: string;
    ContractListID: any;
    siteUrl: any;
    EmployeeDetailListID: any;
    SmartMetaDataListID: any;
}
//# sourceMappingURL=IHrContractSearchProps.d.ts.map